Be less assertive with From and FromName
FromName is now the original text with a note appended indicating
it was sent from the platform.

From is now untouched.

Previously, we overwrote both of these—probably to provide a sense
of consistency from all of the automated emails. It turns out that
this leads to more confusion because various Gravity Forms emails
and other notifications all appear to come from the same place.

This may cause a resurgence in the number of emails that we get as
fake site admins when we don't change the admin email on a site
after handing it over. We'll need to deal with those as they come
up and, when needed, look for more interesting ways of filtering
emails on a case by case basis.